pyproject.toml
[project]
name = "quick-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.0.1",
    "flagembedding>=1.3.4",
    "jupyterlab>=4.4.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-elasticsearch>=0.3.2",
    "langchain-groq>=0.3.2",
    "langchain-ollama>=0.3.2",
    "langchain-openai>=0.3.16",
    "langchain-tavily>=0.1.6",
    "langgraph>=0.4.1",
    "markdownify>=1.1.0",
    "openai>=1.77.0",
    "python-dotenv>=1.1.0",
    "tavily-python>=0.7.2",
]
